+++ lib/money_attribute/migration_extensions/table_definition.rb
@@ -7,14 +7,14 @@ module MoneyAttribute
     module TableDefinition
       include Helper
 
-      def money(accessor, options = {})
+      def money_attribute(accessor, options = {})
         amount_col, currency_col, amount_opts, currency_opts = parse_money_args(accessor, options)
 
         column(amount_col, amount_opts[:type], **amount_opts.except(:type))
         column(currency_col, :string, **currency_opts) if currency_col
       end
 
-      def remove_money(accessor, options = {})
+      def remove_money_attribute(accessor, options = {})
         amount_col, currency_col, = parse_money_args(accessor, options)
 
         remove_column(amount_col)