+++ test/dummy/db/migrate/20240429142109_create_offers.rb
@@ -3,8 +3,7 @@ class CreateOffers < ActiveRecord::Migration[7.1]
     create_table :offers do |t|
       t.string :product
       t.date :date
-      t.decimal :price_amount
-      t.string :price_currency
+      t.money_attribute :price
 
       t.timestamps
     end