+++ test/dummy/db/migrate/20240429184843_create_simple_offers.rb
@@ -3,8 +3,8 @@ class CreateSimpleOffers < ActiveRecord::Migration[7.1]
     create_table :simple_offers do |t|
       t.string :product
       t.date :date
-      t.decimal :price
-      t.decimal :discount
+      t.money_attribute :price, currency: false
+      t.money_attribute :discount, currency: false
 
       t.timestamps
     end