+++ test/dummy/db/migrate/20260602041927_create_financial_transactions.rb
@@ -3,8 +3,7 @@ class CreateFinancialTransactions < ActiveRecord::Migration[8.1]
     create_table :financial_transactions do |t|
       t.string :description
       t.datetime :date
-      t.integer :amount
-      t.string :currency, limit: 3
+      t.money_attribute :amount, amount: { type: :integer }, currency: { limit: 3 }
 
       t.timestamps
     end